Upgrade appengine maven plugin to the cloud-sdk based "2.0" branch. This changes the mvn appengine:* targets:  - mvn appengine:devserver -> appengine:run  - mvn appengine:update -> appengine:deploy
The cloud project ID and version also moves from the appengine-web.xml to the plugin configuration in pom.xml.

Another critical difference is that this plugin only handles the run/deploy phase and does not attempt to package the application. It's thus required to run `mvn package appengine:...` to ensure the .war file is up to date.